refactor: clean up TrainActionManager interface
- Remove ensureForkRoute (navigation, belongs in AutoPilot)
- Remove forceSegmentReset (dead code with TODO body)
- Remove getSavedSpeedBeforeReverse/setSavedSpeedBeforeReverse
  (state moved to Train.java)
- Make internal-only methods private (executeCommand, scheduleResume,
  resumeWaiting, runPendingCommands)
- Move fork-orienting logic from ActionManager to AutoPilotImpl
- Remove dead forceSegmentReset call from AutoPilotImpl.activate()
