+++ src/main/java/letrain/itinerary/impl/AutoPilotImpl.java
@@ -7,11 +7,15 @@ import letrain.itinerary.AutoPilot;
 import letrain.itinerary.Itinerary;
 import letrain.itinerary.SegmentPathfinder;
 import letrain.itinerary.Waypoint;
+import letrain.itinerary.WaypointCommand;
+import letrain.map.Dir;
+import letrain.segments.RailNode;
+import letrain.segments.RailwayGraph;
 import letrain.segments.Segment;
 import letrain.track.rail.ForkRailTrack;
+import letrain.track.rail.RailTrack;
 import letrain.vehicle.rail.impl.Train;
 import letrain.itinerary.TrainActionManager;
-import letrain.itinerary.WaypointCommand;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -163,9 +167,6 @@ public class AutoPilotImpl implements AutoPilot {
         waitTicks = 0;
         pendingCommands.clear();
         currentIndex = 0;
-        if (actionManager != null) {
-            actionManager.forceSegmentReset();
-        }
         log.info("[AP] activate → FOLLOWING");
 
         // Actuación inicial reactiva
@@ -261,9 +262,7 @@ public class AutoPilotImpl implements AutoPilot {
         if (index != -1 && index + 1 < currentRoute.size()) {
             Segment nextSeg = currentRoute.get(index + 1);
             log.info("[AP] onSegmentEntered: orienting fork for next segment {} from {}", nextSeg.getId(), currentSeg.getId());
-            if (actionManager != null) {
-                actionManager.ensureForkRoute(currentSeg, nextSeg);
-            }
+            ensureForkRoute(currentSeg, nextSeg);
         }
     }
 
@@ -285,9 +284,63 @@ public class AutoPilotImpl implements AutoPilot {
 
     @Override
     public void ensureForkRoute(Segment from, Segment to) {
-        if (actionManager != null) {
-            actionManager.ensureForkRoute(from, to);
+        if (train == null || train.getModel() == null)
+            return;
+        RailwayGraph graph = train.getModel().getRailwayGraph();
+        if (graph == null)
+            return;
+
+        var fromSteps = from.getSteps();
+        var toSteps = to.getSteps();
+        if (fromSteps == null || toSteps == null)
+            return;
+
+        var f1 = fromSteps.getFirst();
+        var f2 = fromSteps.getSecond();
+        var t1 = toSteps.getFirst();
+        var t2 = toSteps.getSecond();
+
+        RailNode node = null;
+        if (f1 != null && f1.getRailNode() != null) {
+            var n1 = f1.getRailNode();
+            if ((t1 != null && n1.equals(t1.getRailNode())) || (t2 != null && n1.equals(t2.getRailNode()))) {
+                node = n1;
+            }
         }
+        if (node == null && f2 != null && f2.getRailNode() != null) {
+            var n2 = f2.getRailNode();
+            if ((t1 != null && n2.equals(t1.getRailNode())) || (t2 != null && n2.equals(t2.getRailNode()))) {
+                node = n2;
+            }
+        }
+
+        if (node == null) {
+            log.warn("[AP] ensureForkRoute {}->{}: no shared node found", from.getId(), to.getId());
+            return;
+        }
+        if (!(node.getTrack() instanceof ForkRailTrack fork)) {
+            log.debug("[AP] ensureForkRoute {}->{}: shared node is not a fork ({})", from.getId(), to.getId(),
+                    node.getTrack());
+            return;
+        }
+
+        for (var step : node.getOutSteps()) {
+            Segment nextSeg = graph.getSegment(step);
+            log.debug("[AP] ensureForkRoute {}->{}: outStep dir={} seg={}", from.getId(), to.getId(), step.getDir(),
+                    nextSeg != null ? nextSeg.getId() : "null");
+            if (nextSeg != null && nextSeg.equals(to)) {
+                Dir targetDir = step.getDir();
+                var alt = fork.getRouter().getAlternativeRoute();
+                boolean altNeeded = alt != null && alt.getValue() == targetDir;
+                log.info("[AP] ensureForkRoute {}->{}: MATCH fork={} altNeeded={} currentAlt={}", from.getId(),
+                        to.getId(), fork.getId(), altNeeded, fork.isUsingAlternativeRoute());
+                if (fork.isUsingAlternativeRoute() != altNeeded) {
+                    fork.flipRoute();
+                }
+                return;
+            }
+        }
+        log.warn("[AP] ensureForkRoute {}->{}: no outStep leads to target seg", from.getId(), to.getId());
     }
 
     @Override
